Changing SpinAPI path or board num triggers update
Added value change for SpinAPI path or Board Num to the event structure so
that updating these values in the UI triggers changes properly on `Done'.

Previously it didn't trigger update for either variable, so if only those
values were changed in the dialog the changes would not be propagated to
the device.  This could lead to, for instance, failing to change the board
number and thus connecting to the wrong board as master"