LVTools works with GitHub distro, new instructions
Improvements to tool setup added in commit 0801247:

Powershell packaged with the GitHub distribution neither runs sh by
default nor includes ~/bin/ in the default path.  Thus, changed the
commands to include ~/bin/ as part of their invocation, which should work
also in git-bash.

Switched mergetool key from cmd to command in configuration, as it is
supposed to be.

Put basic instructions in README.md