Add Slave_Pulseblaster Device
Implement a Slave_Pulseblaster device in SetList, including adding it to
the AddDevice and Device Manager parts of SetList.  Functions using wait
steps to hold current output until triggered by master, but this
introduces a delay between trigger edge and output change of 90ns on test
hardware.

If this delay is unnacceptable, a free-running PB could be implemented
that only required an intial trigger and from then on would rely on its
own timing.