Add Slave Channel to Master's Private Data
Prepping for adding ability of slave device to control the output state on
one (or more) of the master's channels by adding a piece of private data.
Using an array of clusters, where each cluster represents control of a
single channel.  Cluster has a numeric element given channel index and an
array of booleans indicating state.

The array will be ordered in the same way as Ramp List (reverse-time
order) to keep processing consistent.

Changes were made in a strict TypeDef, so they propogated out to all
necessary locales.