Move some error checking into Master
Because we presumably want _every_ master to check for conflicts between
slave and trigger channels, that functionality was moved to the master.
Since the master knows nothing about reserved lines in its children, it
outputs the channels it expects to be controlled.  This is
Check_Control_Conflicts.vi

Checking whether a controlled line is reserved is done in
Check_Reserved_Lines.vi, and it takes the list of controlled channels and
compares it against what channels were reserved.

FinalizeHWI.vi was updated with this error-checking done first.