Fixed timing display issues
actual dt is time/npnts, not what Lua uses to calculate OR necessarily
what is in the "step" column, unless time/step is an integer.
For correct plotting of final step in a ramp, need to double the final
point in the waveform to get interpolation to draw to total time.