Add device `Slave Line' for testing purposes
Back-end for Slave Line generated to enable testing of Slave Line
interface.  Is a 1-column device that only does slave control of a single
master channel.  There is no reason to use this for anything but testing,
so it will be removed from the UI (though the code may remain) before
merging to master.