Add default picture to device (Fixes #39)
Uses a '?' as the default picture for devices in case child classes don't
implement their own.  Decided to leave the thrown error if picture not
found as this is an actual error that needs to be fixed, and there is not
appropriate error indicator to connect the VI. The user has the option to
press continue and ignore the problem, so it won't prevent the use of
SetList even if the PNG file somehow gets deleted.

Fixes #39.