Add allowed value checking to MenuCol
If the column does not accept variables as entries, it should only allow
internal contents to contain valid menu entries.  To this end, in both
SetCol.vi and ValChange.vi, if the column does not allow variables, check
contents against list of menu entries.

If not found in the list, present a one-button dialog warning the user
and do not change the value.

If column is a variable _OR_ the entry is valid, make the parent call to
update the table.