Change how new classes are added to AddDevice menu
Now we place a class constant and a string for the menu text into
codes/DeviceVIs/Utils_devs/ImplementedDevices/ImplementedDevices.vi.

The Add Device menu is built at runtime from the contents of said case
structure with a call to UpdateAddDeviceMenu.vi.

AddDevice.vi now call GetChosenDevice.vi to grab the device referenced.

codes/TypeDefs/AddDeviceRing.ctl is switched to back to a non-strict
TypeDef so that UppdateAddDeviceMenu.vi is able to modify the contents at
runtime.