Add `Needs 2nd Pass' to EnableCell, InitAppearance
Because objects that need to modify other cell's enabled/disabled states
with need to do so on a second pass when the table gets processed left to
right, we need to flag when if we want to be included in said second pass.

Added a bool to do this to EnableCell & InitAppearnce, and set to False
for every case _except_ MenuDisableCol's.