Add bool for and implement 2nd pass of InitAppear.
Give InitAppearance.vi a `Needs 2nd Pass' boolean that flags whether or
not it will need a second pass after the rest of the table has been
initialized.  Implemented this in the place it is called
(ApplyTableAppearance_benchmarking.vi).

Also implemented a 2nd pass for specifically MenuDisableCol in
ApplyTableBlockAppearance.vi

Almost done with this bugfix, but for some reason when doing copy/paste
related actions the enabling/disabling gets undone on the whichever row
the click happens.