Add "Insert Function" menu to right-click
"Insert Function" menu parses the User Defined Functions [UDF] field for
all function templates, generates a set of tags/items for them, and takes
care of building & inserting a menu item where appropriate.

Implementing the function menu as a class allows it to process the UDFs
only once and keep things static until further changes are made. This
keeps the overhead related to processing the UDF string small.

Needed to add a reference to the class to the MenuDo, which involved
changing the SetListRefs typedef, which touched every copy of MenuDo.vi

Moved both "Insert Variable" and "Insert Function" related VIs to the
codes/Utils/InsertMenu/ directory to help keep things tidy.