Generate compileable AeroBasic scripts
Fixed some syntax errors in CommandToScript and ParametersToScript. Also
changed some formatting/whitespace to make the output AeroBasic script
easier to read.

The beginning of Load Hardware is in: the generated script is output to
the temporary directory and the path is saved to pass on to the Aerotech
VIs to do the actual programming, etc.