Tabs & SetList table resize with window if running
Set a minimum size for the window to keep from clipping any controls on
the tabs.

At startup will resize things to fit nicely in the window.

Used the VI Panel resize event to trigger more resizing whenever the
window size is changed.  Updates a little slowly during the event, but
should be fine because you can't be doing anything else while this happens
anyway.