Add preferences backend, implement window pos'n
Add SetListPrefs.lvclass, which works towards #22.  Implements a
preference system that is saved to disk in an xml file when the quit
button is pressed, and loaded at startup.  There are VIs available to
force sections to wait until initiliazation of the preferences from disk
is done, allowing you to act on startup from values stored on disk.

Used this system to implement storing a preferred window position that
will automatically be returned to on startup. (If desired).