Add pref. to hide devices in "Add Device" menu
Added preference item to SetListPrefs.lvclass, interface on the
Preferences pane, and the action to the actual menu internals.

Now AddDevice.vi looks up devices by name instead of menu index, but I
left the old functionality in place under a Disable structure, so if this
breaks behavior I wasn't aware of it should be pretty staightforward to
switch it out.

Addresses #22.