Save version from version.txt (Fix #62)
Instead of hard-coding the version number inside SetList.vi, save it in
version.txt so we can avoid changes to SetList.vi when all we've changed
is the version number.  Also makes an obvious place to change/check what
version of SetList you are currently running.

Fixes #62, provided we remember to update this file when we tag.