Remove KeyDown event for faster typing (Fix #52)
Fixes #52 (Slow Typing).

The SetList Table control previous had a KeyDown? filter event attached to
it, which meant that every key press with the table in focus had to be
processed by the main VI.  This resulted in a noticeable slow down in
typing for large tables.  Removing this event speeds up typing, at the
cost of two things:
  - <Return> no longer gets changed to <Enter>, so it now begins editing
    the cell below it
  - <Tab>    no longer gets changed to <Enter>, so it now begins editing
    the cells to its right.

The main snafu is this allows typing outside the normal bounds of the
table.  This will be programmatically removed, but the user may not like
this behavior.

Added a label on the front panel outlining how to stop editing the table.