Clear NI Card if Error on Load (Fixes #68)
DAQmx would report an error if you tried to command out-of-bounds values,
presumably during the Load process.  Errors at that point kept anything
else from running, including the Clear commands that normally occur after
the cycle.  For most devices this seems to be fine, but the NI cards
appear to need their prior tasks cleared before you can load new ones.

@jqi-sr-lab got us pointed down the right direction, now at the tail end
of the Load VI we run a Clear if and only if there was an error.  The
error will still be reported.

This should fix #68.