Process TCP-based Variables commands
Process most commands and place results in appropriate queues, but so far
nothing is actually done with those queues.  That is next on the
implementation.  All commands read in as expected, and the few that aren't
supported respond as such.

CC (Close Connection) command is missing, need to decide the best way to
handle that particular case.