Fix TCP Timeout defaults (0->25s) and error clear
Defaults were at 0s for timeout, resulting in no command ever being
accepted.  Fixed this for all functions that read in values from the TCP
connection.

Also, timeouts were handled properly by the initial command reader, but
were being cleared by subsequent call to TCP Flush inside the command
handler's "BC" case.  Fixed this by fencing it with a Error-No Error case
structure at the input, so the timeout will never be cleared if there is
already a timeout when it is called.