Process TCP-generated variable sets
codes/Utils/ExternalFeedback/LoadVarSetIntoVM.vi provides the interface
for the individual variable sets to get updated in the variable manager.

SetList.vi was modified to incorporate both the "Update Now" options (via
a while loop on a short delay as in the Mulligan functionality) and the
"Update after Sequence" option (via a persistent queue that is checked for
items once the Sequence History Table is filled with "Complete").

Currently will not automatically start a sequence if it was idle and a set
is added for "after sequence".

Currently will not look at the "After Sequence" queue at all if in cycle
mode, I don't think.