Cap TableHistory depth at 100 actions (Fixes #69)
Since each individual change to the table adds a entry to the TableHistory
array, individual tables could grow quite large.  If the same dimensioned
table was left open for a long time without a sufficiently different load
operation, the history length could grow quite large (>1,000 items), and
for already large tables this takes up a lot of memory.  This makes
operations changing the table length expensive, and seems to have been
slowing down making changes to the table (like changing digital values).

To fix this, SaveToTblHist.vi now caps the total history depth to 100.
The number was chosen as a comprimise to history depth and reasonable
resource consumption, erring on the side of more depth than necessary.

This should fix #69.  The spirit of the fix was confirmed by manual
clearing of the table history on a previously slow setup.