Var. Feedback Sequence queue used in "Cycle" Mode
Running in "Cycle" mode (unsequenced, but starting a new cycle at the
conclusion of the previous one) now pulls entries from the sequence queue
and updates the HWI as necessary.  This should enable using "cycle" mode
with external feedback to do automatic optimization routines.

In cycle mode, sending something in the UpdateNow mode results in the
values being updated in the Variable Manager, but because there is no
automated Update HWI action they won't be used in the actual hardware
until either the user manually presses Update HWI -OR- a sequence feedback
is processed (which neccessarily does an Update HWI)