Add ability to request final state for any line
Resolves #57.

Defaults to unchanged from previous line, which is the prior behavior.

When setting this via Update Line Final State.vi, it checks whether or not
that line's final state has already been set to an output of a different
kind.

The options are High, Low, or Same, which are self-explanatory, and also
Default, which should not be used by the user.  Default has the behavior
of Same, but is used internally to distinguish from an entry explicitly
set to Same vs. something unchanged from the default.