Stricter checking variable name in VariableManager
Found strange behavior when including spaces in variable names,
experimentation revealed similar behavior when including traditional
operator symbols (+ - / etc.).  Looking into this showed that Lua
(lua.org/manual see the "Lexical Conventions" section) expects object
names/identifiers to be:
"any string of letters, digits and underscores, not beginning with a
digit"
It also present a list of reserved keywords.

Added the list of reserved keywords to GlobalReservedNames.vi which had
only been checking for names SetList uses internally.

We were already checking for beginning the name with a number, so I only
added a step for checking the name only contains a-z A-Z 0-9 _ (which
happens to be the PCRE \w group).

I also added a case for checking for _[A-Z]+ which according to the same
section are supposed to be reserved for internal Lua use by convention.

Finally, string constant fields containing the empty string were replaced
with the empty string constant for clarity.