Add demo of pre-processing for connected vars
Issue #63 mentions adding the ability for variables to depend on other
variables.  Looking into this functionality  we realized a bunch of things
mentioned in the discussion on GitHub.  One of these is that you have to
be careful because execution order become important.

To demo a way to process variables that depend on each other I created a
toy LabView project that processes functions for what named objects they
depend on, checks whether or not any of those are circular definitions,
and if there are no closed cycles outputs a valid execution order.

For now we've decided not to implement this in SetList itself, but leaving
this with the project may prove useful to some intrepid future programmer
looking to add the functionality.