Inform Igor string gen. after seq. var. update
Sr. Lab reported that sometime Igor would receive incorrect value for
variables compared to what actually happened in the cycle.  This was
likely because the Inform Igor commands are constructed using the
"LastAcquireValues" property, and so if the Igor commands are generated
before LV updates the variable values the commands will be based on the
previous values.

This was fixed by adding Error in/out terminals to Construct Seqquence
Variable Igor String.vi, forking the error terminal on the output of
GetVariableValues.vi, and re-merging the errors after the UpdateHWI vi.
This should ensure that the Igor commands are only generated after the
correct values are obtained.  It has the added benefit that any problems
generating the commands will also be reported to the user as an error.