Table history saved as delimited string (Fix #69)
Actual changes were made in how the Table history TypeDefs were stored,
this cascaded out to many other VIs because it caused changes to the
Reference types used by those VIs.

Table history used to be saved as an array of clusters of 2D string
arrays.  Now we've switched to a 1D array of flat strings, with each
individual string being a full table state.  Table states are stored as
"Spreadsheet Strings" with null characters (ASCII 0x00) as column
delimiters and end-of-line symbols as row delimiters.  Neither of these
are typable in the actual table, so this should be safe.

Testing shows a decrease from average SaveHist time of 40-60ms to around
1ms, which makes a noticeable improvement in responsiveness from LV.  This
time is now also an order of magnitude below that spent actually executing
code in the DigitalColumn:LeftClick.vi, so further gains will likely be
made there.