Fix ISO8601Timestamp.vi for positive TZ offsets
'%z' was properly ouputting a leading '-' for negative timezone (TZ)
offsets, but for positive ones dropped the sign entirely.  This meant that
for US TZ's we tested under everthing looked file, while positive offsets
(Europe, much of Africa, Asia, Australia) had no sign and a trailing ':'.

Fixed this by taking just the '%z' part and prepending a '+' to any string
that lead with a digit instead of a sign.

Closes #76.