Add 'Save to File'; 'Load Recent' works w/ files
Added a 'Save to File' button for those who can't and/or won't utilize the
Git-based save system.

For maximum usefulness, reworked the 'Load Recent' menu to interact well
with both saving _and_ loading from either the Git-based system or from
normal files.  The Git system will continue to populate menu entries with
the formate
<hash> -=- <date saved> -=- <title>
while the file-only based system will do entries as
From File: <path to file>

On load, it checks for the leading string "From File: " (which will never
appear as a hash) and invokes the load from file system if it is found.
Otherwise it will load from the repository.

Closes #75