Handle 'Resume' before 'Stop Expt.' is processed
Stop Expt. (SE) is only processed when deciding whether or not to run the
next item in a sequence. Resume is processed as soon as it is pressed.

The 'resume' logic only handled disabling the SE button, but not setting
it to any particular state.  It also issues a Sequence Event, which
triggers various things, among them whether or not to stop the experiment.
So the sequence event gets processed immediately, sees the SE is true,
sets the SE to false and disables the control. It expects this to be
sufficient to stop.

A Sequence Event is _also_ triggered at the end of a shot. When that
happens, it checks SE, sees that it is false, and continues the sequence,
but there is no code to re-enable the SE button.  This means you are now
stuck running the entire cycle.

To fix, do two things:
 1) Resume buttom sets SE to false. This shouldn't be confusing to the
 user because if you're resuming you shouldn't also expect to be stopping.
 2) Ensure that the SE state is toggled _before_ sending the sequence
 event, to make sure the sequence event sees the appropriate state to act
 on.

Fixes #81 probably. Can't test on this rig.