Prevent Viewer from making calls to Lua
Issue #79 seems to be caused by the Loading routine clearing Lua.
@jiraphat2200 posted a workaround that showed Lua being forced to
re-evaluate the loaded function table resolves things.  Sure enough,
LoadedCluster_toSetList_v5.vi makes a call to close Lua, which is the
source of the problem.

Added a boolean input to said VI which controls whether or not this
closing call is made.  Defaults to true, which should preserve current
behavior anywhere else it is called.

Fixes #79.