Diff with last saved or loaded, not master
This involves checking against the contents of <Loaded Hash> or whatever
it is called.  If this is <FromFile> it will always recommend saving,
because it won't have the diff generated to check against.

Updated ShowDiff.vi to search for ---, +++, and @@ to extract what it
needs from the diff header.  This eased the ability to handle <FromFile>
cases as well.