Send Last Save Location to Igor
Adds LastSave= key to the IgorInform string, and fills it with either full
commit hash (if last save was to a repository) or the full path (If the
last save was to a file) as the value.  This should enable at least easily
finding the files that the data was taken with.

Closes #74.