Dynamically search for and load device classes
Before, we were floating around a `ImplementedDevices.vi` which was
providing a way to index device class names and class defaults. It was
also a source of frequent merge conflicts, because any addition or removal
of a device class required going going into that VI and changing a
manually ordered block diagram.

To fix this, we scrape the `DeviceVIs` directory for subdirectories
containing a file `deviceName.txt` that also contain a `*.lvclass`. If
found, the text file's only line should be the display name for the
device, and the lvclass file can be loaded dynamically from disk.

This also allowed a small amount of block diagram cleanup/optimization in
`Manage Devices.vi` and sub-vis.