DMD code bug fix
The DMD hardware VIs were not reading out the correct number of bytes
from the TCP/IP port for error checking. This was causing the error
checking parts of the VIs to report spurious errors when the DMD was
functioning properly. I've changed the number of read out bytes to 7
(from 6), which eliminates the issue.

DSB