Fix bug in DMD Pattern Sequence, add program mode column
Fixed a couple bugs in the number of bytes written to the DMD in order
to do DMD pattern sequences. Also added a program mode column to the DMD
devices that will let users select whether image(s) should be uploaded
to the DMD every cycle or only if there are changes to the image path.