Add preference for IP for shared variable deploy
Previously this was done via a configuration in SetList.lvproj to override
the use of localhost.  This was causing some problems as commits from some
users overwrote these configurations, changing what they were for other
lab's control computers.

To fix, we've turned off auto-deployment and switched to programmatic
deployment.  IP is configurable via a preference.  If left blank this
keeps LVs behavior, whatever that may be.

Gracefully fails if you put in a bad IP by returning the field to the
previously working IP.

Although a Undeploy vi has been added, we never call it, because if you
change the IP from one of two NICs to the others you might end in a state
where nothing is deployed (this is bad, as it breaks communication between
the control and acquisition computer).

Fixes #93