Fix PCRE `(\s|\S)`->`(?s).` to fix missed pref matches
We were occasionally not matching the preference section for the Load
Recent menu.  This was causing its ModificationTime field to remain empty,
while simultaneously writing the field at the end of the file.  This broke
the XML parser, throwing an error on preference load.

It seems to have been called by a problem with the PCRE matching.  The LV
docs mentioned using `(?s).` as the match for any character including
newlines, so tried switching to that instead of `(\s|\S)` (which is match
any whitespace or non-whitespace character).  This seems to have corrected
the issue, at least local to the two testing machines I have access to.

Fixes #96.