Add warning dialog for failed git interactions
Errors with git were causing 'Load from repo' to silently fail, never
displaying either the front panel or a dialog indicating an error to the
user.

Added simple error handlers to the three dialogs a user is likely
to encounter git-related errors. They include the error message if git is
configured incorrectly.

This does result in a bit of error box spam in some cases, but compared to
silent failure this is the way to go.

Fixes #102